Добавляет файлы mocks.ts, types/types.ts и переносит const.ts в src
