+++ Notes.md
@@ -13,9 +13,18 @@
 
 ----
 **TO DO**
-- Добавить проверку активной вкладки в компоненте `location-item` и добавлять класс `tabs__item--active` - это видимо на этапе роутинга
+- Добавить проверку активной вкладки в компоненте `location-item` и добавлять класс `tabs__item--active` - это видимо на этапе роутинга ( с помощью компонента NavLink в Router)
 - Вынести в отдельный компонент блок сортировки
 - Маршруты можно вынести в const.js в перечеслинение enum
-- Подсветку активной вкладки можно будет сделать с помощью компонента NavLink в Router
+- Моки вынети в отдельный файл + пример моков взять из ТЗ
+- Более тщательное разбиение на компоненты
+- Типы можно вынести в src/types/offers.ts или types.ts
+
+- const.tsx переименовать в const.ts и перенести в src/
+- моки не в const.ts, а в моки вынести
+- в PlaceCard для пропсов отдельный тип сделать PlaceCardProps
+- компонент футера сделать
+- plasecList.length передавать вместо константы
+
