+++ src/components/locations/location-list.tsx
@@ -1,5 +1,5 @@
 import LocationItem from './location-item';
-import { LocationType } from '../const';
+import { LocationType } from '../../const';
 
 type LocationListProps = {
   locations: LocationType[];
