+++ src/components/places/place-list.tsx
@@ -1,5 +1,5 @@
 import PlaceCard from './place-card';
-import { PlaceType } from '../const';
+import { PlaceType } from '../../const';
 
 type PlaceListProps = {
   places: PlaceType[];
