+++ src/const.tsx
@@ -1,0 +1,15 @@
+export const PLACES_COUNT: number = 312;
+
+export enum AppRoute {
+  Root = '/',
+  Favorites = '/favorites',
+  Login = '/login',
+  Offer = '/offer',
+  AnyOther = '*'
+}
+
+export enum AutorisationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN'
+}
