+++ src/types/types.ts
@@ -1,0 +1,16 @@
+export type PlaceType = {
+  title: string;
+  price: string;
+  type: string;
+  description: string;
+  link: string;
+  isPremium: boolean;
+  isInFavorite: boolean;
+  rating: number;
+  imgSrc: string;
+};
+
+export type LocationType = {
+  title: string;
+  link: string;
+};
